replace fasbf to fabs.

+++ JasperMobile/Sources/ViewControllers/Repository&Libraries/ResourceViewer/JMResourceViewerViewController/DashboardViewer/JMDashboardLoader/JMVisualizeDashboardLoader/JMVisDashboardLoader.m
@@ -126,7 +126,7 @@ typedef NS_ENUM(NSInteger, JMDashboardViewerAlertViewType) {
 
 - (void)updateViewportScaleFactorWithValue:(CGFloat)scaleFactor
 {
-    if ( fabsf(self.visualizeManager.viewportScaleFactor - scaleFactor) >= 0.49 ) {
+    if ( fabs(self.visualizeManager.viewportScaleFactor - scaleFactor) >= 0.49 ) {
         self.visualizeManager.viewportScaleFactor = scaleFactor;
 
         JMJavascriptRequest *request = [JMJavascriptRequest new];
